Added URL functionality
Added a text box on the main download page.  Type a URL in the box and hit the download button to download a video.  Should work as long as the element id and the input remains as text of some kind.  Feel free to mess with the CSS or different input types to make it look nice/fit in.